#4: Added new FileUtility class. #4: Removed premature optimization of ServiceDefinitionsRepository indexing. #4: Moved TestServiceDefinitionScannerProvider to a mock package in test sources.
